requirements.txt

fastapi>=0.95.0
uvicorn[standard]>=0.18.0
numpy>=1.24.0
openai>=1.37.0
agno>=1.1.6
python-dotenv>=1.0.1
faiss-cpu>=1.8.0
pydantic>=2.4.0
jinja2>=3.1.2
httpx>=0.24.0
